fix: enforce dp and use-only in tests

+++ examples/f90/issue_2639_intrinsic_subroutine_calls_ok.f90
@@ -1,8 +1,9 @@
 program issue_2639_intrinsic_subroutine_calls_ok
+    use, intrinsic :: iso_fortran_env, only: dp => real64
     implicit none
 
-    real :: t
-    real :: r
+    real(dp) :: t
+    real(dp) :: r
     integer :: count
     integer :: rate
     integer :: max_count
